pycaffe/advanced/data-augmentation-testing/requirements.txt
caffe
missinglink-sdk
opencv-python